A bloackchain Application
CertiCheck
A blockchain application to make the certification process easier for the HR interviewer. It constructs a Merkle Tree which contains nodes consisting of the hashes generated by the data given by the universities. The root of this Merkle tree is the combined hash of the all individual hashes, this root is then put on the blockchain. When the applicant uploads his resume, it parses the resume and extracts significant details like Name, CPI, College Name and Year of Graduation using NLP tecniques and stores it in a json format and creates a hash of that and combine it with the neighboring hashes and generates the root hash. It then verifies this root hash with the hash on the blockchain and accordingly decides the validity of the resume.